Handle bungled input
Apparently the AVS team can call intents in the certification process
that ignores slot rules. I've not been able to reproduce this via the
test tab nor my Echo, but their rejection notice indicated this is
possible. So if our intent is triggered and a number isn't passed or
our parseInt results in NaN, bail.